+++ example/.eslintrc.js
@@ -1,0 +1,16 @@
+module.exports = {
+  root: true,
+  extends: '@react-native-community',
+  rules: {
+    'prettier/prettier': [
+      'error',
+      {
+        quoteProps: 'consistent',
+        singleQuote: true,
+        tabWidth: 2,
+        trailingComma: 'es5',
+        useTabs: false,
+      },
+    ],
+  },
+};