+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] - 2024-02-28
+
+- iOS SDK version:  6.8.0
+- Android SDK version: 14.0.1
+
+### React Native
+
+#### Fixed
+
+- Resolved compatibility issue on Android with RN 0.77+
+
 ## [3.13.0] - 2024-12-20
 
 - iOS SDK version:  6.6.3